requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
APScheduler==3.10.4
paho-mqtt==2.1.0
Jinja2==3.1.4
email-validator==2.2.0
